yourlocalchef/requirements.txt
django==3.0.3
djangorestframework==3.10.3
django-filter==2.0.0
mock==2.0.0
Pillow==2.1.0